feat(ui): extract audit log list to a component
