+++ modules/front-end/src/app/features/safe/audit-logs/index/index.component.less
@@ -1,52 +1,1 @@
-@import "variables";
-
-section.body-container {
-  width: 100%;
-  height: 100%;
-  padding-top: 20px;
-  padding-left: 36px;
-
-  nz-input-group {
-    width: 250px;
-  }
-
-  .table-content-area {
-    height: 100%;
-  }
-
-  .audit-logs {
-    height: calc(100% - 85px);
-    overflow: auto;
-    padding: 0 24px;
-    margin-top: 40px;
-
-    nz-empty {
-      margin-top: 40px;
-    }
-  }
-
-  .audit-logs-group {
-    .group-title {
-      font-weight: 700;
-      font-size:16px
-    }
-
-    .group-items {
-      display: flex;
-      flex-direction: column;
-      gap: 8px;
-      background-color: @white;
-      margin-top: 15px;
-      border-radius: 10px;
-      padding: 0 10px;
-    }
-  }
-
-  .load-more {
-    margin-top: -15px;
-    text-align: center;
-  }
-}
-
-
