--- modules/front-end/src/app/features/safe/feature-flags/details/audit-logs/audit-logs.component.less
@@ -1,47 +1,0 @@
-@import "variables";
-
-.body-container {
-  width: 100%;
-  height: 100%;
-  padding-left: 24px;
-
-  nz-input-group {
-    width: 250px;
-  }
-
-  .table-content-area {
-    height: 100%;
-  }
-  .audit-logs {
-    height: calc(100% - 85px);
-    overflow: auto;
-    padding: 0 24px;
-    margin-top: 40px;
-
-    nz-empty {
-      margin-top: 40px;
-    }
-  }
-
-  .audit-logs-group {
-    .group-title {
-      font-weight: 700;
-      font-size:16px
-    }
-
-    .group-items {
-      display: flex;
-      flex-direction: column;
-      gap: 8px;
-      background-color: @white;
-      margin-top: 15px;
-      border-radius: 10px;
-      padding: 0 10px;
-    }
-  }
-
-  .load-more {
-    margin-top: -15px;
-    text-align: center;
-  }
-}
